fix: revisi soal 2a dan 2b

+++ soal2/soal2a.c
@@ -10,23 +10,23 @@
 void main()
 {
     int i, j, k;
-    int size_m1[2], size_m2[2];
+    long long size_m1[2], size_m2[2];
 
     // matriks 1 [row = 4, col = 3]
     size_m1[0] = 4; 
     size_m1[1] = 3; 
-    int m1[size_m1[0]][size_m1[1]];
+    long long m1[size_m1[0]][size_m1[1]];
 
     // matriks 2 [row = 3, col = 6]
     size_m2[0] = 3; 
     size_m2[1] = 6; 
-    int m2[size_m2[0]][size_m2[1]];
+    long long m2[size_m2[0]][size_m2[1]];
 
     for (i = 0; i < size_m1[0]; i++)
     {
         for (j = 0; j < size_m1[1]; j++)
         {
-            scanf("%d", &m1[i][j]);
+            scanf("%lld", &m1[i][j]);
         }
     }
 
@@ -34,12 +34,12 @@ void main()
     {
         for (j = 0; j < size_m2[1]; j++)
         {
-            scanf("%d", &m2[i][j]); 
+            scanf("%lld", &m2[i][j]); 
         }
     }
 
     // perkalian matriks
-    int mresult[size_m1[0]][size_m2[1]];
+    long long mresult[size_m1[0]][size_m2[1]];
     memset(mresult, 0, sizeof(mresult));
     for (i = 0; i < size_m2[0]; i++)
     {
@@ -54,9 +54,9 @@ void main()
 
     // shared memory
     key_t key = 1234;
-    int (* value)[size_m2[1]];
+    long long (* value)[size_m2[1]];
 
-    int shm_id = shmget(key, sizeof(int[size_m1[0]][size_m2[1]]), IPC_CREAT | 0666);
+    int shm_id = shmget(key, sizeof(long long[size_m1[0]][size_m2[1]]), IPC_CREAT | 0666);
     value = shmat(shm_id, NULL, 0);
 
     printf("\nMultiplication Result Matrix : \n");
@@ -65,7 +65,7 @@ void main()
         for (j = 0; j < size_m2[1]; j++)
         {
             value[i][j] = mresult[i][j];
-            printf("%-2d ", value[i][j]);
+            printf("%-2lld ", value[i][j]);
         }
         printf("\n");
     }
